Converter/requirements.txt
coremltools==4.0b2
numpy==1.19.1
Pillow==7.2.0
